Added Character Ability DB Services
Merchants now sell Base Ability types and Ability Events.

Added CharacterAbilityService. Abilities are now saved to the database.
Added CharacterKnownAbilityService. Known abilities are now saved to the database.
Added Character Ability synchronization to the Client upon connect.

+++ FishMMO-Unity/Assets/Scripts/Shared/Entity/Ability/Activation/AbilityActivationReplicateData.cs
@@ -6,10 +6,10 @@ namespace FishMMO.Shared
 	public struct AbilityActivationReplicateData : IReplicateData
 	{
 		public bool InterruptQueued;
-		public int QueuedAbilityID;
+		public long QueuedAbilityID;
 		public KeyCode HeldKey;
 
-		public AbilityActivationReplicateData(bool interruptQueued, int queuedAbilityID, KeyCode heldKey)
+		public AbilityActivationReplicateData(bool interruptQueued, long queuedAbilityID, KeyCode heldKey)
 		{
 			InterruptQueued = interruptQueued;
 			QueuedAbilityID = queuedAbilityID;
